(Debatable) improvements, for details see below
Debatable because I am not sure why the original phrase is chosen. Some
might come from a general name, I changed to something easier to
recognise and understand. Some might be from a specific name, but I
don't know what the pinyin refers to (e.g. Zhuàng Chuán for a warship),
I gave them a more general name.

1. Standardization. The original words are probably bad machine
translations, which is not common words in Chinese.
- Zhuàng Chuán -> Zhàn Chuán (战船, or Méng Chōng 艨艟, used in Han
  period for a common warship)
- Chì Chuán -> Huǒ Chuán (火船, fire ship)
- Yú Pái -> Yú Chuán (渔船, fishing ship)

2. Jiànkè is more like a swordsman by his own, instead of a soldier
in a army. Nowadays it also has the meaning of an "assassin".
- Jiànkè -> Dāo Bīng

3. Unlike "forge", Duànzào only works as a verb. Added a "Chǎng" to
indicate the place for forge.
- Duànzào -> Duànzào Chǎng

4. Mǎ hào is okay for "stable", but not common enough for recognizability.
- Mǎ hào -> Mǎ Jiù (马厩)

5. I don't know what the original word is at all:
- Guān Sì -> Gōngdiàn (宫殿, meaning palace)
  This might deviate from the original word, but to distinguish from
  "imperial ministry/court", see the following item.
- Bùzhǎng jí -> Guān Shǔ (官署, meaning buildings of local government)
- Xuànfēng -> Tóushí Chē (投石车, meaning a trebuchet)
  Xuànfēng means whirlwind/cyclone, the closest thing I have found is
  Pīlì Chē (霹雳车), a name given to the trebuchets of Liú Bāng. Pīlì
  means "thunder", referring to the loud sound trebuchets make.
- Shoushe Nu -> Nǔ Chē (弩车, a crossbow vehicle)
  I haven't found anything with pinyin of Shoushe.